Move MockVWS to a public module path
MockVWS was defined in mock_vws._requests_mock_server.decorators, so the
class's __module__ pointed through a private package even though it is
part of the public API.

Move the module to mock_vws.decorators. The documented
``from mock_vws import MockVWS`` import is unchanged.
