+++ src/mock_vws/__init__.py
@@ -1,8 +1,8 @@
 """Tools for using a fake implementation of Vuforia."""
 
 from mock_vws._mock_common import MissingSchemeError
-from mock_vws._requests_mock_server.decorators import MockVWS
 from mock_vws.cloud_query import CloudQueryFailureResponse
+from mock_vws.decorators import MockVWS
 from mock_vws.model_target import (
     ModelTargetGenerationFailure,
     ModelTargetGenerationWarning,
