+++ src/mock_vws/decorators.py
@@ -13,6 +13,12 @@ from requests import PreparedRequest
 from responses import RequestsMock
 
 from mock_vws._mock_common import MissingSchemeError, RequestData
+from mock_vws._requests_mock_server.mock_web_query_api import (
+    MockVuforiaWebQueryAPI,
+)
+from mock_vws._requests_mock_server.mock_web_services_api import (
+    MockVuforiaWebServicesAPI,
+)
 from mock_vws._respx_mock_server.decorators import start_respx_router
 from mock_vws.cloud_query import CloudQueryFailureResponse
 from mock_vws.database import CloudDatabase, VuMarkDatabase
@@ -30,9 +36,6 @@ from mock_vws.target_raters import (
     TargetTrackingRater,
 )
 from mock_vws.vumark import VuMarkGenerationFailure
-
-from .mock_web_query_api import MockVuforiaWebQueryAPI
-from .mock_web_services_api import MockVuforiaWebServicesAPI
 
 if TYPE_CHECKING:
     import respx
